examples/example-ai-anthropic/pyproject.toml
[project]
name = "example-ai-anthropic"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.80.0",
    "opentelemetry-instrumentation-anthropic>=0.24.0",
    "opentelemetry-sdk>=1.30.0",
    "opentelemetry-exporter-otlp-proto-http>=1.30.0",
]
